voice_diary/pyproject.toml
[project]
name = "voice-diary"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "atsuya", email = "f.atsuya.personal@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "blinker==1.9.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "dnspython==2.7.0",
    "flask==3.1.0",
    "google-generativeai>=0.8.4",
    "itsdangerous==2.2.0",
    "jinja2==3.1.4",
    "markupsafe==3.0.2",
    "pymongo==4.10.1",
    "pyttsx3==2.98",
    "werkzeug==3.1.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voice_diary"]
